Attempt 1: use dragenter, dragleave and move events to implement potential container hint effect
This solution seems to work as a user would expect in all cases. There's only one
visual caveat when a figure, that is already inside a container that is inside
another container, is dragged. In that case, only most inner-most container will glow.
That contradicts the visual hinting when the user might drag another figure from outside
both containers into the inner container, on which event that outer container will
glow too.